Amount to payDebt can be uint128

+++ contracts/interfaces/IMarginModule.sol
@@ -37,7 +37,7 @@ interface IMarginModule is IBasePerpMarket {
      * @notice Pays back debt for `accountId` and `marketId`.
      * Users can partially pay off their debt and if amount is bigger than the debt, only the debt will be cleared.
      */
-    function payDebt(uint128 accountId, uint128 marketId, uint256 amount) external;
+    function payDebt(uint128 accountId, uint128 marketId, uint128 amount) external;
 
     /**
      * @notice Convenience method to withdraw all collateral for `accountId` and `marketId`.
